keylogger/requirements.txt
pynput
pyperclip
Pillow
schedule 